pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "job-tracker"
version = "0.1.0"
description = "Self-hosted hour-tracking tool with local web UI. Bilingual (EN/pt-BR)."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Rafael Krause", email = "rafaelkrause@gmail.com" }]
requires-python = ">=3.10"
keywords = ["time-tracking", "flask", "self-hosted", "productivity"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Flask",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Natural Language :: Portuguese (Brazilian)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Scheduling",
]
dependencies = [
    "flask>=3.0",
    "flask-babel>=4.0",
]

[project.optional-dependencies]
tray = ["pystray>=0.19", "Pillow>=10.0"]
dev = [
    "ruff>=0.6",
    "mypy>=1.11",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pre-commit>=3.8",
    "mkdocs-material>=9.5",
    "mkdocs-static-i18n>=1.2",
    "babel>=2.16",
]

[project.urls]
Homepage = "https://github.com/rafaelkrause/job_tracker"
Documentation = "https://rafaelkrause.github.io/job_tracker/"
Repository = "https://github.com/rafaelkrause/job_tracker"
Issues = "https://github.com/rafaelkrause/job_tracker/issues"
Changelog = "https://github.com/rafaelkrause/job_tracker/blob/main/CHANGELOG.md"

[project.scripts]
job-tracker = "app.__main__:main"

[tool.hatch.build]
# Compiled translation catalogs (`.mo`) are gitignored but must ship inside
# the distributions. `artifacts` opts them back in without disabling
# VCS-based exclusion for everything else.
artifacts = ["app/i18n/**/*.mo"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = [
    "app/",
    "run.py",
    "README.md",
    "LICENSE",
    "CHANGELOG.md",
    "requirements.txt",
]

# --- Ruff (lint + format) ---
[tool.ruff]
target-version = "py310"
line-length = 100
extend-exclude = [".venv", "build", "dist", "site", "app/i18n"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF", "N"]
ignore = ["E501"]  # handled by formatter

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D", "N802"]
"run.py" = ["N"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

# --- mypy ---
[tool.mypy]
python_version = "3.10"
files = ["app"]
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = false   # progressive strictness
check_untyped_defs = true

# --- pytest ---
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config"
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:flask_babel.*",
]

# --- coverage ---
[tool.coverage.run]
source = ["app"]
branch = true
omit = ["app/tray.py", "app/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
show_missing = true
skip_covered = false
